Edit 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,24 +8,24 @@
 int main(void)
 {
 	int i = 0;
-	int a = 1;
-	int b = 2;
+	long int a = 1;
+	long int b = 2;
 
 	while (i < 98)
 	{
 		if (i == 0)
 		{
-			printf("%d", a);
+			printf("%ld", a);
 		}
 		else if (i == 1)
 		{
-			printf(", %d", b);
+			printf(", %ld", b);
 		}
 		else
 		{
-			int f = a + b;
+			long int f = a + b;
 
-			printf(", %d", f);
+			printf(", %ld", f);
 			a = b;
 			b = f;
 		}
